# config/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Watch & Do API",
        "description": "Watch & Do API",
        "contact": {
            "name": "Alexandre Maia",
            "url": "http://wwwalexandremaia.pt",
            "email": "info@alexandremaia.pt"
        },
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/comments": {
            "get": {
                "tags": [ "Comments" ],
                "description": "<h2>Get User Comments List</h2><br><p>Role required: <strong>Register</strong></p>",
                "responses": {
                    "200": {
                        "description": "Comments List",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Comments" },
                                "examples": {
                                    "Favorite List": 
                                    [
                                        {
                                            "id": 2,
                                            "userId": 3,
                                            "videoId": 1,
                                            "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus."
                                        },
                                        {
                                            "id": 4,
                                            "userId": 3,
                                            "videoId": 2,
                                            "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus."
                                          }
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "post": {
                "tags": [ "Comments" ],
                "description": "<h2>Create new Comment</h2><br><p>Users can only create own comments</p><br><p>Role required: <strong>Register</strong></p>",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#components/schemas/Comments" },
                            "examples": {
                                "Favorite": {
                                    "value": {
                                        "videoId": 1,
                                        "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New Comment",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Comments" }
                            },
                            "text/plain": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": { "type": "string" },
                                "examples": {
                                    "Error": { "value": "Error" }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": { "type": "string" },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/comments/{id}": {
            "put": {
                "tags": [ "Comments" ],
                "description": "<h2>Archive own Comment</h2><br>Only Admins or higher can access archived Comments<br><p>Role required: <strong>Register</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Success": {
                                    "value": "Comment Archive"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/comments": {
            "get": {
                "tags": [ "Comments" ],
                "description": "<h2>Get Comments List</h2><br><p>Role required: <strong>Admin</strong></p>",
                "responses": {
                    "200": {
                        "description": "Comments List",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Comments" },
                                "examples": {
                                    "Favorite List": 
                                    [
                                        {
                                            "id": 2,
                                            "userId": 3,
                                            "videoId": 1,
                                            "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus.",
                                            "status": true,
                                            "archived": false,
                                            "createdAt": "2020-12-02T17:42:16.000Z",
                                            "updatedAt": "2020-12-02T17:42:16.000Z"
                                        },
                                        {
                                            "id": 4,
                                            "userId": 3,
                                            "videoId": 2,
                                            "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus.",
                                            "status": true,
                                            "archived": false,
                                            "createdAt": "2020-12-04T08:33:54.000Z",
                                            "updatedAt": "2020-12-04T08:33:54.000Z"
                                          }
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "post": {
                "tags": [ "Comments" ],
                "description": "<h2>Create new Comment</h2><br><p>Role required: <strong>Admin</strong></p>",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#components/schemas/Comments" },
                            "examples": {
                                "Favorite": {
                                    "value": {
                                        "userId": 1,
                                        "videoId": 1,
                                        "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New Comment",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Comments" }
                            },
                            "text/plain": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": { "type": "string" },
                                "examples": {
                                    "Error": { "value": "Error" }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": { "type": "string" },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/comments/relations": {
            "get": {
                "tags": [ "Comments" ],
                "description": "<h2>Get Comment List with related items</h2><br><p>Relations:</p><p>-> User;</p><p>-> Video</p><br><p>Role required: <strong>Admin</strong></p>",
                "responses": {
                    "200": {
                        "description": "Comment List with Relations",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Comments" },
                                "examples": {
                                    "Comments List ": [
                                        {
                                            "id": 4,
                                            "userId": 3,
                                            "videoId": 2,
                                            "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus.",
                                            "status": true,
                                            "archived": false,
                                            "createdAt": "2020-12-04T08:33:54.000Z",
                                            "updatedAt": "2020-12-04T08:33:54.000Z",
                                            "Video": {
                                                "id": 2,
                                                "title": "teste title 2",
                                                "youtubeid": "testeyoutubeid 2",
                                                "status": true,
                                                "createdAt": "2020-12-01T16:14:47.000Z",
                                                "updatedAt": "2020-12-01T16:14:47.000Z"
                                            },
                                            "User": {
                                                "id": 3,
                                                "username": "registered",
                                                "email": "registered@mail.pt",
                                                "password": "$2b$10$yUhKHY3B/lgSEiZjVpz/4OtRxXwY40HtEaW8kCqsdvizMwGgKp9oi",
                                                "status": true,
                                                "role_id": 1,
                                                "createdAt": "2020-11-28T23:43:35.000Z",
                                                "updatedAt": "2020-12-01T09:39:59.000Z"
                                            }
                                        },
                                        {
                                            "id": 6,
                                            "userId": 3,
                                            "videoId": 3,
                                            "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus.",
                                            "status": true,
                                            "archived": false,
                                            "createdAt": "2020-12-04T16:23:50.000Z",
                                            "updatedAt": "2020-12-04T16:23:50.000Z",
                                            "Video": {
                                                "id": 3,
                                                "title": "teste title 3",
                                                "youtubeid": "testeyoutubeid 3",
                                                "status": true,
                                                "createdAt": "2020-12-01T16:51:13.000Z",
                                                "updatedAt": "2020-12-01T16:51:13.000Z"
                                            },
                                            "User": {
                                                "id": 3,
                                                "username": "registered",
                                                "email": "registered@mail.pt",
                                                "password": "$2b$10$yUhKHY3B/lgSEiZjVpz/4OtRxXwY40HtEaW8kCqsdvizMwGgKp9oi",
                                                "status": true,
                                                "role_id": 1,
                                                "createdAt": "2020-11-28T23:43:35.000Z",
                                                "updatedAt": "2020-12-01T09:39:59.000Z"
                                            }
                                        }
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
...
        "/wd/comments/{id}/relations": {
            "get": {
                "tags": [ "Comments" ],
                "description": "<h2>Get Comments info by Id with relations</h2><br><p>Relations:</p><p>-> User;</p><p>-> Video</p><br><p>Role required <strong>Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment Info with relations",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Comments" },
                                "examples": {
                                    "Favorite": {
                                        "id": 4,
                                        "userId": 3,
                                        "videoId": 2,
                                        "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus. Proin est tortor, euismod vitae magna vitae, facilisis lobortis dolor. Aenean tempus venenatis mauris. Ut convallis nibh at est bibendum, mollis mattis ligula faucibus.",
                                        "status": true,
                                        "archived": false,
                                        "createdAt": "2020-12-04T08:33:54.000Z",
                                        "updatedAt": "2020-12-04T08:33:54.000Z",
                                        "Video": {
                                            "id": 2,
                                            "title": "teste title 2",
                                            "youtubeid": "testeyoutubeid 2",
                                            "status": true,
                                            "createdAt": "2020-12-01T16:14:47.000Z",
                                            "updatedAt": "2020-12-01T16:14:47.000Z"
                                        },
                                        "User": {
                                            "id": 3,
                                            "username": "registered",
                                            "email": "registered@mail.pt",
                                            "password": "$2b$10$yUhKHY3B/lgSEiZjVpz/4OtRxXwY40HtEaW8kCqsdvizMwGgKp9oi",
                                            "status": true,
                                            "role_id": 1,
                                            "createdAt": "2020-11-28T23:43:35.000Z",
                                            "updatedAt": "2020-12-01T09:39:59.000Z"
                                        }
                                    }
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
...
    }
}